backport from Batocera38

+++ input/drivers/dos_input.c
@@ -49,7 +49,7 @@ static void dos_keyboard_free(void)
          dos_key_state[i][j] = 0;
 }
 
-static int16_t dos_input_state(
+static int32_t dos_input_state(
       void *data,
       const input_device_driver_t *joypad,
       const input_device_driver_t *sec_joypad,
@@ -70,7 +70,7 @@ static int16_t dos_input_state(
          if (id == RETRO_DEVICE_ID_JOYPAD_MASK)
          {
             unsigned i;
-            int16_t ret = 0;
+            int32_t ret = 0;
 
             for (i = 0; i < RARCH_FIRST_CUSTOM_BIND; i++)
             {
