+++ input/drivers_joypad/ctr_joypad.c
@@ -118,13 +118,13 @@ static int16_t ctr_joypad_axis(unsigned port_num, uint32_t joyaxis)
    return ctr_joypad_axis_state(port_num, joyaxis);
 }
 
-static int16_t ctr_joypad_state(
+static int32_t ctr_joypad_state(
       rarch_joypad_info_t *joypad_info,
       const struct retro_keybind *binds,
       unsigned port)
 {
    unsigned i;
-   int16_t ret                          = 0;
+   int32_t ret                          = 0;
    uint16_t port_idx                    = joypad_info->joy_idx;
 
    if (port_idx >= DEFAULT_MAX_PADS)
