+++ input/drivers_joypad/ps2_joypad.c
@@ -132,13 +132,13 @@ static int16_t ps2_joypad_axis_state(unsigned port_num, uint32_t joyaxis)
    return val;
 }
 
-static int16_t ps2_joypad_state(
+static int32_t ps2_joypad_state(
       rarch_joypad_info_t *joypad_info,
       const struct retro_keybind *binds,
       unsigned port)
 {
    unsigned i;
-   int16_t ret                          = 0;
+   int32_t ret                          = 0;
    uint16_t port_idx                    = joypad_info->joy_idx;
 
    if (port_idx >= DEFAULT_MAX_PADS)
