+++ input/drivers_joypad/wiiu/kpad_driver.c
@@ -114,13 +114,13 @@ static int16_t kpad_axis(unsigned port, uint32_t axis)
          data.is_negative);
 }
 
-static int16_t kpad_state(
+static int32_t kpad_state(
       rarch_joypad_info_t *joypad_info,
       const struct retro_keybind *binds,
       unsigned port)
 {
    unsigned i;
-   int16_t ret                          = 0;
+   int32_t ret                          = 0;
    uint16_t port_idx                    = joypad_info->joy_idx;
 
    for (i = 0; i < RARCH_FIRST_CUSTOM_BIND; i++)
