+++ network/netplay/netplay.h
@@ -264,7 +264,7 @@ void netplay_load_savestate(netplay_t *netplay,
  **/
 void netplay_core_reset(netplay_t *netplay);
 
-int16_t netplay_input_state(netplay_t *netplay,
+int32_t netplay_input_state(netplay_t *netplay,
       unsigned port, unsigned device,
       unsigned idx, unsigned id);
 
@@ -337,7 +337,7 @@ void video_frame_net(const void *data, unsigned width,
       unsigned height, size_t pitch);
 void audio_sample_net(int16_t left, int16_t right);
 size_t audio_sample_batch_net(const int16_t *data, size_t frames);
-int16_t input_state_net(unsigned port, unsigned device,
+int32_t input_state_net(unsigned port, unsigned device,
       unsigned idx, unsigned id);
 
 #ifdef HAVE_NETPLAYDISCOVERY
